An attempt at using SCOOP to parallelize the thermo data estimation.
Run as python -m scoop -vv rmg.py path/to/input.py

If this works, it will be slow, as we're pickling and passing
the database every time we try to evaluate a species.

However, it's worse than that because it seems the database is 
unpickleable, giving a "TypeError: can't pickle instancemethod objects".

Debugging SCOOP with Python 2.x sucks, because it hides where 
the error is (apparently this is better in Python 3). However,
this same error can be generated by
        import pickle
        print "trying to pickle the database"
        print pickle.dumps(database)