Get multiprocessing working for thermo estimation.
Estimating the thermo of a species updates not only
the species.thermo but also the species.conformer.
This must also be preserved.

Also, storing the pool of workers in the CoreEdgeReactionModel
instance means it can't be pickled (for Restart) so we now
store the pool as a module level variable.