Additional resource logging when running in a PBS script.
If running inside a PBS script then 'ja' should print resource usage
info for the current job based on
 http://svn.nics.tennessee.edu/repos/pbstools/trunk/bin/ja

Unfortunately running this command repeatedly via Popen creates a new fork each
time, causing memory problems of its own. Instead we make a single subprocess,
if we appear to be in a PBS environment, and keep it running.
I also do some processing of the PBS job resource information
and add it to the statistics.xls spreadsheet.

If $PBS_JOBID is not set, this should have no effect.