Change the way in which libraryOrder is handled for kinetics database.
Prior to this, libraryOrder listed all the libraries used in order of priority,
but did not indicate whether they are Seed mechs or reaction libraries.

Now, indicate this by setting libraryOrder as a list of tuples in the format
('library_label', LibraryType), where LibraryType is set to either 'Reaction Library'
or 'Seed'.  That way we can avoid trying to double generating reactions for the 'Seed'
library type (since they have already been added to the core).  This also gives us
more flexibility for treating the seeds and reaction libraries differently.