Keeping the mopac output file verification consistent with gaussian
Checks were added in gaussian.py for cases where the InChI did not fully print.
This rarely happens in mopac but the check should still exist.

Implementation of the geometry conversion to graph has been modified to match that in
gaussian.py. There is no change to the algorithm, just modified it to be consistent.