Change installation instructions back.
We have now copied the rdkit osx binary to our channel,
so we no longer need two channels, which seemed to be causing
problems on linux. I suspect we are probably having caching issues,
so our testing may not be great. But fingers crossed.
These instructions now seem to work for both Connie (linux) and me (OS X)