Add a SimulationProfilePlotter class which plots the top species as the simulation is running.
Turned on only if rmg job is saving simulation profiles.
It adds a little bit of CPU time, but in my opinion
is a worthy trade off.

It is possible to directly write from object to graph instead of
parsing from csv, but it is simpler to do it this way
so that the csv files are the basis for other types of plots.