Add averageThermoData() function to rmgpy.data.thermo
This function is to be used for scripts designed to fit groups or do
other things with thermo data from libraries.  But it can be extended to do other
interesting things in the future.

Average a list of thermoData values together.
Sets uncertainty values to be the approximately the 95% confidence interval, equivalent to
2 standard deviations calculated using the sample standard variance:

Uncertainty = 2s
s = sqrt( sum(abs(x - x.mean())^2) / N - 1) where N is the number of values averaged

Note that uncertainties are only computed when number of values is greater than 1.