Fix appending of library reactions in CoreEdgeReactionModel.addReactionLibraryToOutput()
Fix bug that occurred because of bad identification of the reaction library
label.  Also no longer try to create new LibraryReaction objects,
because they are already in memory.

Fix the saveChemkin function so that when saving edge, it does not save the
outputspeciesList or outputreactionlist, because that would print a bunch of duplicates.

Also no longer check duplicates, because it is performed in enlarge() for
all core and edge reactions