Unit test: TestReactionToCantera()
Tests that we create the correct Cantera reaction objects and that the units conversion
work correctly.  More unit tests may be necessary before we can fully trust the conversion, but
for the most part this should cover almost all the things we care about.

We might be more worried with using highly varied units, however, we anticipate that
most frequently this function will be used after a model is finished and reloading through
a chemkin file, in which case the most generic units are typically in use.

The Arrhenius.toCanteraKinetics() class also will catch unexpected units and raise an exception
at the moment, so that is a good thing.