Increasing accuracy of some chemkin output.
This is to help the conversion to cantera models
in the model comparing scripts of the importer,
but wouldn't hurt in general.

Mostly they're increasing activation energies to
0.0001 kcal/mol (i.e. 0.1 cal/mol) which is more
precise than needed but helped consistency with
an imported model.

The other change that's worth preserving is changing
the pressure accuracy in PLOG formats, which I think
was an issue in RMG-Py.