Made a jupyter notebook to discover low T pathways.
The goal of this notebook is to allow you to generate all
the species that would occur by systematically applying
known low-temperature pathways to a starting fuel molecule.

You might do this to generate species that you then put
into an RMG input file.
Such an input file block is generated at the end of the notebook.