Read all Chebyshev coefficients from Chemkin files
Some chemkin files, like those published with
https://doi.org/10.1016/j.combustflame.2019.04.040.
put some of their Chebyshev coefficients on the first 
line after declaring the order of the polynomials.

RMG used to assume that the first line only contains
the orders, and all the coefficients appeared on subsequent
lines. This commit fixes that assumption, and also adds 
a check that the right number of coefficients have been read.