Import Species from species (to avoid Julia)
I think we just need the Species class which
we can import directly from rmgpy.species
instead of from rmgpy.rmg.model which
in turn imports it from rmgpy.species
(but which also imports a bunch of other stuff while loading).
This is part of a general campaign to reduce un-needed import dependencies
and perhaps make it so some partial imports
can avoid loading all the RMS julia stuff.