SMILES from OpenBabel now replace `[Pt]` with `*`
This means they can be parsed in a round trip by RDKit
(the default SMILES reader). 
This is handy because OpenBabel is the default SMILES
*writer* for things with an N atom, but not everything.
Now it's more consistent, outputting a * for a surface site.

I added a unit test for round-trip conversion to and from 
SMILES a few times for various adsorbates including some with N.